Simplified JPEG transcoding pipeline to only operate on floats and in the last step 32-bit ints
Benchmark from BenchmarkDotNet:
```
Scalar

| Method                  | Mean     | Error    | StdDev   | Gen0     | Gen1     | Gen2     | Allocated  |
|------------------------ |---------:|---------:|---------:|---------:|---------:|---------:|-----------:|
| ShortRowTranscode       | 80.71 ms | 0.775 ms | 0.725 ms | 142.8571 | 142.8571 | 142.8571 | 1137.88 KB |
| FloatRowTranscode       | 66.37 ms | 0.971 ms | 0.909 ms | 250.0000 | 250.0000 | 250.0000 | 1138.88 KB |
| ShortFastTranscode  | 64.57 ms | 0.337 ms | 0.315 ms | 125.0000 | 125.0000 | 125.0000 |  915.77 KB |
| FloatFastTranscode  | 47.73 ms | 0.350 ms | 0.311 ms | 181.8182 | 181.8182 | 181.8182 |  947.58 KB |

AVX

| Method                  | Mean     | Error    | StdDev   | Gen0     | Gen1     | Gen2     | Allocated  |
|------------------------ |---------:|---------:|---------:|---------:|---------:|---------:|-----------:|
| ShortRowTranscode       | 47.79 ms | 0.212 ms | 0.198 ms | 181.8182 | 181.8182 | 181.8182 |  1137.9 KB |
| FloatRowTranscode       | 44.86 ms | 0.331 ms | 0.309 ms | 272.7273 | 272.7273 | 272.7273 | 1138.87 KB |
| ShortFastTranscode  | 21.08 ms | 0.071 ms | 0.060 ms | 187.5000 | 187.5000 | 187.5000 |  915.79 KB |
| FloatFastTranscode  | 20.52 ms | 0.142 ms | 0.126 ms | 187.5000 | 187.5000 | 187.5000 |  947.65 KB |
```

+++ src/PdfToSvg/Imaging/Jpeg/JpegBitmap.cs
@@ -16,7 +16,7 @@ namespace PdfToSvg.Imaging.Jpeg
 
         public static JpegBitmap Empty { get; } = new JpegBitmap(0, 0, 0);
 
-        public readonly short[] Data;
+        public readonly float[] Data;
         public readonly int Width;
         public readonly int Height;
         public readonly int Components;
@@ -27,7 +27,7 @@ namespace PdfToSvg.Imaging.Jpeg
             Width = width;
             Height = height;
             Components = components;
-            Data = new short[width * height * components];
+            Data = new float[width * height * components];
         }
 
         public void DrawNearestNeighbourClippedOnto(
@@ -57,7 +57,7 @@ namespace PdfToSvg.Imaging.Jpeg
             }
         }
 
-        public void GetBlock(short[] block,
+        public void GetBlock(float[] block,
             int x, int y, int componentIndex,
             int subSamplingX, int subSamplingY)
         {
@@ -79,7 +79,7 @@ namespace PdfToSvg.Imaging.Jpeg
                     ((y + iy * subSamplingY) * Width + x) * Components +
                     componentIndex;
 
-                var lastValue = (short)0;
+                var lastValue = 0f;
 
                 for (var ix = 0; ix < maxX; ix++)
                 {
